SVM
SVM:
creates a cross-validated model for a SVM, achieves similar accuracy to the KNN model. Same style as KNN model.

KNN: I think you meant 6 for n_neighbors instead of 5. The highest f1 score is when k=6, not k=5. I had an error on my pc which was fixed by adding ".astype("U")" to df.content